refactor: detail 추가, 삭제 컴포넌트 분리
